next commit

+++ controller/LikeController.js
@@ -1,43 +1,65 @@
+const ensureAuthorization = require('../auth'); // 인증 모듈
+const jwt = require('jsonwebtoken');
 const conn = require('../mariadb'); // db 모듈
 const {StatusCodes} = require('http-status-codes'); // status code 모듈
 
 const addLike = (req, res) => {
     // 좋아요 추가
+    const book_id = req.params.id;
 
-    const {id} = req.params; // book_id
-    const {user_id} = req.body;
+    let authorization = ensureAuthorization(req, res);
 
-    let sql = "INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?);";
-    let values = [user_id, id];
-    conn.query(sql, values,
-        (err, results) => {
-            if(err) {
-                console.log(err);
-                return res.status(StatusCodes.BAD_REQUEST).end();
-            }
+    if(authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError){
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "잘못된 토큰 입니다."
+        });
+    } else {
+        let sql = "INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?);";
+        let values = [authorization.id, book_id];
+        conn.query(sql, values,
+            (err, results) => {
+                if(err) {
+                    console.log(err);
+                    return res.status(StatusCodes.BAD_REQUEST).end();
+                }
 
-            return res.status(StatusCodes.OK).json(results);
-    })
-}
+                return res.status(StatusCodes.OK).json(results);
+        })
+    }
+};
 
 const removeLike = (req, res) => {
     // 좋아요 제거(취소)
-    const {id} = req.params; // book_id
-    const {user_id} = req.body;
+    const book_id = req.params.id;
 
-    let sql = "DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?;";
-    let values = [user_id, id];
-    conn.query(sql, values,
-        (err, results) => {
-            if(err) {
-                console.log(err);
-                return res.status(StatusCodes.BAD_REQUEST).end();
-            }
+    let authorization = ensureAuthorization(req, res);
 
-            return res.status(StatusCodes.OK).json(results);
-    })
+    if(authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError){
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "잘못된 토큰 입니다."
+        });
+    } else {
+        let sql = "DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?;";
+        let values = [authorization.id, book_id];
+        conn.query(sql, values,
+            (err, results) => {
+                if(err) {
+                    console.log(err);
+                    return res.status(StatusCodes.BAD_REQUEST).end();
+                }
 
-}
+                return res.status(StatusCodes.OK).json(results);
+        })
+    }
+};
 
 module.exports = {
     addLike,
